move prices assertion from common context
